requirements.txt
asgiref==3.4.1
Django==3.2.5
django-mysql==3.12.0
djangorestframework==3.12.4
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
